I #788: Moved LFRicArgDescriptor in its own class
