+++ src/psyclone/dynamo0p3.py
@@ -48,13 +48,13 @@ import os
 from enum import Enum
 from collections import OrderedDict, namedtuple
 import fparser
-from psyclone.parse.kernel import Descriptor, KernelType, getkerneldescriptors
+from psyclone.parse.kernel import KernelType, getkerneldescriptors
 from psyclone.parse.utils import ParseError
-import psyclone.expression as expr
 from psyclone import psyGen
 from psyclone.configuration import Config
 from psyclone.core.access_type import AccessType
 from psyclone.domain.lfric import FunctionSpace
+from psyclone.domain.lfric import LFRicArgDescriptor
 from psyclone.psyir.nodes import Loop, Literal, Schedule
 from psyclone.errors import GenerationError, InternalError, FieldNotFoundError
 from psyclone.psyGen import PSy, Invokes, Invoke, InvokeSchedule, \
@@ -89,14 +89,7 @@ QUADRATURE_TYPE_MAP = {
                            "type": "quadrature_edge_type",
                            "proxy_type": "quadrature_edge_proxy_type"}}
 
-# ---------- API datatypes (scalars, fields, operators) --------------------- #
-GH_VALID_SCALAR_NAMES = ["gh_real", "gh_integer"]
-GH_VALID_FIELD_NAMES = ["gh_field"]
-GH_VALID_OPERATOR_NAMES = ["gh_operator", "gh_columnwise_operator"]
-GH_VALID_ARG_TYPE_NAMES = GH_VALID_FIELD_NAMES + GH_VALID_OPERATOR_NAMES + \
-    GH_VALID_SCALAR_NAMES
-
-# ---------- Fortran datatypes ------------------------- #
+# ---------- Fortran datatypes ---------------------------------------------- #
 SUPPORTED_FORTRAN_DATATYPES = ["real", "integer", "logical"]
 
 # ---------- Stencils ------------------------------------------------------- #
@@ -162,7 +155,7 @@ VALID_LOOP_TYPES = ["dofs", "colours", "colour", ""]
 # ---------- psyGen mappings ------------------------------------------------ #
 # Mappings used by non-API-Specific code in psyGen
 psyGen.MAPPING_SCALARS = {"iscalar": "gh_integer", "rscalar": "gh_real"}
-psyGen.VALID_ARG_TYPE_NAMES = GH_VALID_ARG_TYPE_NAMES
+psyGen.VALID_ARG_TYPE_NAMES = LFRicArgDescriptor.VALID_ARG_TYPE_NAMES
 
 # ---------- Functions ------------------------------------------------------ #
 
@@ -291,539 +284,6 @@ class DynFuncDescriptor03(object):
         for idx, arg in enumerate(self._operator_names):
             res += "  operator_name[{0}] = '{1}'".format(idx+1, arg) + \
                    os.linesep
-        return res
-
-
-class LFRicArgDescriptor(Descriptor):
-    '''
-    This class captures the information specified in one of LFRic API argument
-    descriptors (scalars, fields and operators).
-
-    :param arg_type: Dynamo0.3 argument type (scalar, field or operator)
-    :type arg_type: :py:class:`psyclone.expression.FunctionVar` or \
-                    :py:class:`psyclone.expression.BinaryOperator`
-
-    :raises ParseError: if a 'meta_arg' entry is not of 'arg_type' type.
-    :raises ParseError: if a 'meta_arg' entry has fewer than 2 args.
-    :raises ParseError: if an argument type is not one of LFRic (Dynamo0.3) \
-                        API valid argument types (scalar, field, operator).
-    :raises ParseError: if a field vector notation is not in the correct \
-                        format '(field*n)' where 'n' is an integer.
-    :raises ParseError: if the field vector notation is used for the vector \
-                        size of less than 1.
-    :raises ParseError: if the field vector notation uses a wrong separator \
-                        operator.
-    :raises ParseError: if the field vector notation is used for an \
-                        argument that is not a field.
-    :raises InternalError: if the argument type checks fail when an invalid \
-                           argument type is passed in.
-    :raises ParseError: if the second 'meta_arg' entry is not a valid \
-                        access descriptor.
-    :raises ParseError: if an argument that is not a real scalar has a \
-                        reduction access.
-    :raises InternalError: if none of the checks caught an invalid argument.
-
-    '''
-    def __init__(self, arg_type):
-        self._arg_type = arg_type
-        # Initialise properties
-        self._type = None
-        self._function_space_to = None
-        self._function_space_from = None
-        self._function_space = None
-        self._function_spaces = []
-        # Set vector size to 1 (scalars set it to 0 in their validation)
-        self._vector_size = 1
-        # Initialise other internal arguments
-        self._access_type = None
-        self._function_space1 = None
-        self._function_space2 = None
-        self._stencil = None
-        self._mesh = None
-
-        # Check for correct type descriptor
-        if arg_type.name != 'arg_type':
-            raise ParseError(
-                "In the Dynamo0.3 API each 'meta_arg' entry must be of type "
-                "'arg_type', but found '{0}'".format(arg_type.name))
-
-        # We require at least 2 args
-        if len(arg_type.args) < 2:
-            raise ParseError(
-                "In the Dynamo0.3 API each 'meta_arg' entry must have at "
-                "least 2 args, but found '{0}'".format(len(arg_type.args)))
-
-        # Check the first argument descriptor. If it is a binary operator
-        # then it has to be a field vector with an "*n" appended. If it is
-        # a variable then it can be other allowed type of argument.
-        if isinstance(arg_type.args[0], expr.BinaryOperator):
-            # We expect 'field_type * n' to have been specified
-            argtype = arg_type.args[0].toks[0].name
-            operator = arg_type.args[0].toks[1]
-            # First check for a valid argument type...
-            if argtype not in GH_VALID_ARG_TYPE_NAMES:
-                raise ParseError(
-                    "In the Dynamo0.3 API the 1st argument of a 'meta_arg' "
-                    "entry should be a valid argument type (one of {0}), but "
-                    "found '{1}' in '{2}'".format(GH_VALID_ARG_TYPE_NAMES,
-                                                  self._type, arg_type))
-            # ... and set it if the check passes
-            self._type = argtype
-
-            # Now try to find the vector size for a field vector and return
-            # an error if it is not an integer number...
-            try:
-                vectsize = int(arg_type.args[0].toks[2])
-            except TypeError:
-                raise ParseError(
-                    "In the Dynamo0.3 API field vector notation expects the "
-                    "format '(field*n)' where 'n' is an integer, but the "
-                    "following '{0}' was found in '{1}'.".
-                    format(str(arg_type.args[0].toks[2]), arg_type))
-            # ... or it is less than 2 (1 is the default for all fields)...
-            if vectsize < 2:
-                raise ParseError(
-                    "In the Dynamo0.3 API the 1st argument of a 'meta_arg' "
-                    "entry may be a vector but if so must contain a valid "
-                    "integer vector size in the format '(field*n, "
-                    "where n > 1)', but found n = {0} in '{1}'".
-                    format(vectsize, arg_type))
-            # ... and set the vector size if all checks pass
-            self._vector_size = vectsize
-
-            # Check that the separator operator is correct
-            if operator != "*":
-                raise ParseError(
-                    "In the Dynamo0.3 API the 1st argument of a 'meta_arg' "
-                    "entry may be a vector but if so must use '*' as the "
-                    "separator in the format '(field*n)', but found '{0}' "
-                    "in '{1}'".format(operator, arg_type))
-
-            # Check than no other arguments than fields use vector notation
-            if self._type not in GH_VALID_FIELD_NAMES and self._vector_size:
-                raise ParseError(
-                    "In the Dynamo0.3 API vector notation is only supported "
-                    "for a {0} argument type but found '{1}'".
-                    format(GH_VALID_FIELD_NAMES, arg_type.args[0]))
-
-        elif isinstance(arg_type.args[0], expr.FunctionVar):
-            # We expect 'arg_type' to have been specified
-            if arg_type.args[0].name not in GH_VALID_ARG_TYPE_NAMES:
-                raise ParseError(
-                    "In the Dynamo0.3 API the 1st argument of a 'meta_arg' "
-                    "entry should be a valid argument type (one of "
-                    "{0}), but found '{1}' in '{2}'".
-                    format(GH_VALID_ARG_TYPE_NAMES, arg_type.args[0].name,
-                           arg_type))
-            self._type = arg_type.args[0].name
-        else:
-            raise InternalError(
-                "LFRicArgDescriptor.__init__(): invalid argument type after "
-                "checks, should not get to here")
-
-        # The 2nd arg is an access descriptor (TODO: 3rd in case of scalar)
-        # Convert from GH_* names to the generic access type:
-        api_config = Config.get().api_conf("dynamo0.3")
-        access_mapping = api_config.get_access_mapping()
-        try:
-            self._access_type = access_mapping[arg_type.args[1].name]
-        except KeyError:
-            valid_names = api_config.get_valid_accesses_api()
-            raise ParseError(
-                "In the Dynamo0.3 API the 2nd argument of a 'meta_arg' entry "
-                "must be a valid access descriptor (one of {0}), but found "
-                "'{1}' in '{2}'".format(valid_names,
-                                        arg_type.args[1].name, arg_type))
-
-        # Reduction access descriptors are only valid for real scalar arguments
-        if self._type != "gh_real" and self._access_type in \
-           AccessType.get_valid_reduction_modes():
-            raise ParseError(
-                "In the Dynamo0.3 API a reduction access '{0}' is only valid "
-                "with a real scalar argument, but '{1}' was found".
-                format(self._access_type.api_specific_name(),
-                       self._type))
-
-        # FIELD, OPERATOR and SCALAR argument types descriptors and checks
-        # Fields
-        if self._type in GH_VALID_FIELD_NAMES:
-            self._validate_field(arg_type)
-
-        # Operators
-        elif self._type in GH_VALID_OPERATOR_NAMES:
-            self._validate_operator(arg_type)
-
-        # Scalars
-        elif self._type in GH_VALID_SCALAR_NAMES:
-            self._validate_scalar(arg_type)
-
-        # We should never get to here (#TODO: check if tested)
-        else:
-            raise InternalError(
-                "LFRicArgDescriptor.__init__(): failed argument validation, "
-                "should not get to here")
-
-        # Initialise the parent class
-        super(LFRicArgDescriptor,
-              self).__init__(self._access_type, self._function_space1,
-                             stencil=self._stencil, mesh=self._mesh)
-
-    def _validate_field(self, arg_type):
-        '''
-        Validates argument descriptors for field arguments and
-        populates argument properties accordingly.
-
-        :param arg_type: Dynamo0.3 field (vector) argument type
-        :type arg_type: :py:class:`psyclone.expression.FunctionVar` or \
-                        :py:class:`psyclone.expression.BinaryOperator`
-
-        :raises InternalError: if argument type other than a field is \
-                               passed in.
-        :raises ParseError: if there are fewer than 3 metadata arguments.
-        :raises ParseError: if there are more than 4 metadata arguments.
-        :raises ParseError: if the 3rd argument is not one of valid \
-                            function spaces.
-        :raises ParseError: if the optional 4th argument is not a stencil \
-                            specification or a mesh identifier (for \
-                            inter-grid kernels).
-        :raises ParseError: if a field on a discontinuous function space \
-                            has 'gh_inc' access.
-        :raises ParseError: if a field on a continuous function space has \
-                            'gh_readwrite' access.
-        :raises ParseError: if a field on 'any_space' function space has \
-                            'gh_readwrite' access.
-        :raises ParseError: if a field stencil argument is not read-only.
-
-        '''
-        # Check whether something other than a field is passed in
-        if self._type not in GH_VALID_FIELD_NAMES:
-            raise InternalError(
-                "LFRicArgDescriptor._validate_field(): expecting a field "
-                "argument but got an argument of type '{0}'. Should be "
-                "impossible.".format(arg_type.args[0]))
-
-        # There must be at least 3 arguments
-        if len(arg_type.args) < 3:
-            raise ParseError(
-                "In the Dynamo0.3 API each 'meta_arg' entry must have at "
-                "least 3 arguments if its first argument is of a {0} type, "
-                "but found {1} in '{2}'".format(GH_VALID_FIELD_NAMES,
-                                                len(arg_type.args), arg_type))
-        # There must be at most 4 arguments
-        if len(arg_type.args) > 4:
-            raise ParseError(
-                "In the Dynamo0.3 API each 'meta_arg' entry must have at "
-                "most 4 arguments if its first argument is of a {0} type, "
-                "but found {1} in '{2}'".format(GH_VALID_FIELD_NAMES,
-                                                len(arg_type.args), arg_type))
-
-        # The 3rd argument must be a valid function space name
-        if arg_type.args[2].name not in \
-           FunctionSpace.VALID_FUNCTION_SPACE_NAMES:
-            raise ParseError(
-                "In the Dynamo0.3 API the 3rd argument of a 'meta_arg' "
-                "entry must be a valid function space name (one of {0}) if "
-                "its first argument is of a {1} type, but found '{2}' in "
-                "'{3}".format(FunctionSpace.VALID_FUNCTION_SPACE_NAMES,
-                              GH_VALID_FIELD_NAMES,
-                              arg_type.args[2].name, arg_type))
-        self._function_space1 = arg_type.args[2].name
-
-        # The optional 4th argument is either a stencil specification
-        # or a mesh identifier (for inter-grid kernels)
-        if len(arg_type.args) == 4:
-            try:
-                from psyclone.parse.kernel import get_stencil, get_mesh
-                if "stencil" in str(arg_type.args[3]):
-                    self._stencil = get_stencil(arg_type.args[3],
-                                                VALID_STENCIL_TYPES)
-                elif "mesh" in str(arg_type.args[3]):
-                    self._mesh = get_mesh(arg_type.args[3], VALID_MESH_TYPES)
-                else:
-                    raise ParseError("Unrecognised metadata entry")
-            except ParseError as err:
-                raise ParseError(
-                    "In the Dynamo0.3 API the 4th argument of a 'meta_arg' "
-                    "field entry must be either a valid stencil specification"
-                    "or a mesh identifier (for inter-grid kernels). However, "
-                    "entry '{0}' raised the following error: {1}".
-                    format(arg_type, str(err)))
-
-        # Test allowed accesses for fields
-        if self._function_space1.lower() in \
-           FunctionSpace.VALID_DISCONTINUOUS_NAMES \
-           and self._access_type == AccessType.INC:
-            raise ParseError(
-                "It does not make sense for a field on a discontinuous "
-                "space ('{0}') to have a 'gh_inc' access".
-                format(self._function_space1.lower()))
-        if self._function_space1.lower() in \
-           FunctionSpace.CONTINUOUS_FUNCTION_SPACES \
-           and self._access_type == AccessType.READWRITE:
-            raise ParseError(
-                "It does not make sense for a field on a continuous "
-                "space ('{0}') to have a 'gh_readwrite' access".
-                format(self._function_space1.lower()))
-        # TODO: extend restriction to "gh_write" for kernels that loop
-        # over cells (issue #138) and update access rules for kernels
-        # (built-ins) that loop over DoFs to accesses for discontinuous
-        # quantities (issue #471)
-        if self._function_space1.lower() in \
-           FunctionSpace.VALID_ANY_SPACE_NAMES \
-           and self._access_type == AccessType.READWRITE:
-            raise ParseError(
-                "In the Dynamo0.3 API a field on 'any_space' cannot have "
-                "'gh_readwrite' access because it is treated as continuous")
-        if self._stencil and self._access_type != AccessType.READ:
-            raise ParseError("A stencil must be read-only so its access "
-                             "should be 'gh_read'")
-
-    def _validate_operator(self, arg_type):
-        '''
-        Validates argument descriptors for operator arguments and
-        populates argument properties accordingly.
-
-        :param arg_type: Dynamo0.3 operator argument type
-        :type arg_type: :py:class:`psyclone.expression.FunctionVar`
-
-        :raises InternalError: if argument type other than an operator is \
-                               passed in.
-        :raises ParseError: if there are not exactly 4 metadata arguments.
-        :raises ParseError: if the function space to- is not one of the \
-                            valid function spaces.
-        :raises ParseError: if the function space from- is not one of the \
-                            valid function spaces.
-        :raises ParseError: if the operator argument has an invalid access.
-
-        '''
-        # Check whether something other than an operator is passed in
-        if self._type not in GH_VALID_OPERATOR_NAMES:
-            raise InternalError(
-                "LFRicArgDescriptor._validate_operator(): expecting an "
-                "operator argument but got an argument of type '{0}'. "
-                "Should be impossible.".format(self._type))
-
-        # We expect 4 arguments with the 3rd and 4th each being a
-        # function space
-        if len(arg_type.args) != 4:
-            raise ParseError(
-                "In the Dynamo0.3 API each 'meta_arg' entry must have 4 "
-                "arguments if its first argument is one of {0}, but found "
-                "{1} in '{2}'".
-                format(GH_VALID_OPERATOR_NAMES, len(arg_type.args), arg_type))
-
-        # Operator arguments need to have valid to- and from- function spaces
-        if arg_type.args[2].name not in \
-           FunctionSpace.VALID_FUNCTION_SPACE_NAMES:
-            raise ParseError(
-                "In the Dynamo0.3 API the 3rd argument of a 'meta_arg' "
-                "operator entry must be a valid function space name (one of "
-                "{0}), but found '{1}' in '{2}".
-                format(FunctionSpace.VALID_FUNCTION_SPACE_NAMES,
-                       arg_type.args[2].name, arg_type))
-        self._function_space1 = arg_type.args[2].name
-        if arg_type.args[3].name not in \
-           FunctionSpace.VALID_FUNCTION_SPACE_NAMES:
-            raise ParseError(
-                "In the Dynamo0.3 API the 4th argument of a 'meta_arg' "
-                "operator entry must be a valid function space name (one of "
-                "{0}), but found '{1}' in '{2}".
-                format(FunctionSpace.VALID_FUNCTION_SPACE_NAMES,
-                       arg_type.args[3].name, arg_type))
-        self._function_space2 = arg_type.args[3].name
-
-        # Test allowed accesses for operators
-        if self._access_type == AccessType.INC:
-            raise ParseError(
-                "In the Dynamo0.3 API operators cannot have a 'gh_inc' "
-                "access because they behave as discontinuous quantities")
-
-    def _validate_scalar(self, arg_type):
-        '''
-        Validates argument descriptors for scalar arguments and
-        populates argument properties accordingly.
-
-        :param arg_type: Dynamo0.3 scalar argument type
-        :type arg_type: :py:class:`psyclone.expression.FunctionVar`
-
-        :raises InternalError: if argument type other than a scalar is \
-                               passed in.
-        :raises ParseError: if there are not exactly 2 metadata arguments.
-        :raises ParseError: if scalar arguments do not have a read-only or
-                            a reduction access.
-
-        '''
-        # Check whether something other than a scalar is passed in
-        if self._type not in GH_VALID_SCALAR_NAMES:
-            raise InternalError(
-                "LFRicArgDescriptor._validate_scalar(): expecting a scalar "
-                "argument but got an argument of type '{0}'. Should be "
-                "impossible.".format(arg_type.args[0]))
-
-        # There must be at least 2 arguments to describe a scalar
-        if len(arg_type.args) != 2:
-            raise ParseError(
-                "In the Dynamo0.3 API each 'meta_arg' entry must have 2 "
-                "arguments if its first argument is 'gh_{{r,i}}scalar', but "
-                "found {0} in '{1}'".format(len(arg_type.args), arg_type))
-
-        # Test allowed accesses for scalars (read_only or reduction)
-        if self._access_type not in [AccessType.READ] + \
-           AccessType.get_valid_reduction_modes():
-            api_config = Config.get().api_conf("dynamo0.3")
-            rev_access_mapping = api_config.get_reverse_access_mapping()
-            api_specific_name = rev_access_mapping[self._access_type]
-            valid_reductions = AccessType.get_valid_reduction_names()
-            raise ParseError(
-                "In the Dynamo0.3 API scalar arguments must be "
-                "read-only ('gh_read') or a reduction ({0}) but found "
-                "'{1}' in '{2}'".format(valid_reductions, api_specific_name,
-                                        arg_type))
-
-        # Scalars don't have vector size
-        self._vector_size = 0
-
-    @property
-    def type(self):
-        '''
-        Returns the type of the argument (gh_field, gh_operator, ...).
-
-        :returns: type of the argument.
-        :rtype: str
-
-        '''
-        return self._type
-
-    @property
-    def function_space_to(self):
-        '''
-        Returns the "to" function space for an operator. This is
-        the first function space specified in the metadata.
-
-        :returns: "to" function space for an operator.
-        :rtype: str
-
-        :raises RuntimeError: if this is not an operator.
-
-        '''
-        if self._type in GH_VALID_OPERATOR_NAMES:
-            return self._function_space1
-        raise RuntimeError(
-            "In the Dynamo0.3 API 'function_space_to' only makes sense "
-            "for one of {0}, but this is a '{1}'".
-            format(GH_VALID_OPERATOR_NAMES, self._type))
-
-    @property
-    def function_space_from(self):
-        '''
-        Returns the "from" function space for an operator. This is
-        the second function space specified in the metadata.
-
-        :returns: "from" function space for an operator.
-        :rtype: str
-
-        :raises RuntimeError: if this is not an operator.
-
-        '''
-        if self._type in GH_VALID_OPERATOR_NAMES:
-            return self._function_space2
-        raise RuntimeError(
-            "In the Dynamo0.3 API 'function_space_from' only makes sense "
-            "for one of {0}, but this is a '{1}'".
-            format(GH_VALID_OPERATOR_NAMES, self._type))
-
-    @property
-    def function_space(self):
-        '''
-        Returns the function space name that this instance operates on
-        depending on the argument type: a single function space for a field,
-        function_space_from for an operator and nothing for a scalar.
-
-        :returns: function space than an argument instance operates on.
-        :rtype: str
-
-        :raises InternalError: if an invalid argument type is passed in.
-
-        '''
-        if self._type in GH_VALID_FIELD_NAMES:
-            return self._function_space1
-        if self._type in GH_VALID_OPERATOR_NAMES:
-            return self._function_space2
-        if self._type in GH_VALID_SCALAR_NAMES:
-            return None
-        raise InternalError("LFRicArgDescriptor.function_space(), should "
-                            "not get to here.")
-
-    @property
-    def function_spaces(self):
-        '''
-        Returns the function space names that this instance operates on as
-        a list depending on the argument type: one function space for a
-        field, both function spaces for an operator and an empty list for
-        a scalar.
-
-        :returns: function spaces than an argument instance operates on.
-        :rtype: list of str
-
-        :raises InternalError: if an invalid argument type is passed in.
-
-        '''
-        if self._type in GH_VALID_FIELD_NAMES:
-            return [self.function_space]
-        if self._type in GH_VALID_OPERATOR_NAMES:
-            # return to before from to maintain expected ordering
-            return [self.function_space_to, self.function_space_from]
-        if self._type in GH_VALID_SCALAR_NAMES:
-            return []
-        raise InternalError("LFRicArgDescriptor.function_spaces(), should "
-                            "not get to here.")
-
-    @property
-    def vector_size(self):
-        '''
-        Returns the vector size of the argument. This will be 1 if *n
-        has not been specified for all arguments except scalars (their
-        vector size is set to 0).
-
-        :returns: vector size of the argument.
-        :rtype: str
-
-        '''
-        return self._vector_size
-
-    def __str__(self):
-        '''
-        Creates a string representation of the argument descriptor. This
-        is type and access for scalars with the addition of function
-        space(s) for fields and operators.
-
-        :returns: string representation of the argument descriptor.
-        :rtype: str
-
-        :raises InternalError: if an invalid argument type is passed in.
-
-        '''
-        res = "LFRicArgDescriptor object" + os.linesep
-        res += "  argument_type[0]='{0}'".format(self._type)
-        if self._vector_size > 1:
-            res += "*"+str(self._vector_size)
-        res += os.linesep
-        res += "  access_descriptor[1]='{0}'"\
-               .format(self._access_type.api_specific_name())\
-               + os.linesep
-        if self._type in GH_VALID_FIELD_NAMES:
-            res += "  function_space[2]='{0}'".format(self._function_space1) \
-                   + os.linesep
-        elif self._type in GH_VALID_OPERATOR_NAMES:
-            res += "  function_space_to[2]='{0}'".\
-                   format(self._function_space1) + os.linesep
-            res += "  function_space_from[3]='{0}'".\
-                   format(self._function_space2) + os.linesep
-        elif self._type in GH_VALID_SCALAR_NAMES:
-            pass  # we have nothing to add if we're a scalar
-        else:  # we should never get to here
-            raise InternalError("LFRicArgDescriptor.__str__(), should not "
-                                "get to here.")
         return res
 
 
@@ -1123,7 +583,7 @@ class DynKernMetadata(KernelType):
             if arg.access != AccessType.READ:
                 write_count += 1
                 # We must not write to a field on a read-only function space
-                if arg.type in GH_VALID_FIELD_NAMES and \
+                if arg.type in LFRicArgDescriptor.VALID_FIELD_NAMES and \
                    arg.function_spaces[0] in \
                    FunctionSpace.READ_ONLY_FUNCTION_SPACES:
                     raise ParseError(
@@ -1134,7 +594,7 @@ class DynKernMetadata(KernelType):
                 # We must not write to scalar arguments if it's not a
                 # built-in
                 if self.name not in BUILTIN_MAP and \
-                   arg.type in GH_VALID_SCALAR_NAMES:
+                   arg.type in LFRicArgDescriptor.VALID_SCALAR_NAMES:
                     raise ParseError(
                         "A user-supplied Dynamo 0.3 kernel must not "
                         "write/update a scalar argument but kernel {0} has "
@@ -1213,7 +673,7 @@ class DynKernMetadata(KernelType):
         non_field_arg_types = set()
         for arg in self._arg_descriptors:
             # Collect info so that we can check inter-grid kernels
-            if arg.type in GH_VALID_FIELD_NAMES:
+            if arg.type in LFRicArgDescriptor.VALID_FIELD_NAMES:
                 if arg.mesh:
                     # Argument has a mesh associated with it so this must
                     # be an inter-grid kernel
@@ -1332,13 +792,15 @@ class DynKernMetadata(KernelType):
                     "two fields) but kernel {0} has {1} arguments".
                     format(self.name, len(self._arg_descriptors)))
             # Check that the other two arguments are fields
-            farg_read = psyGen.args_filter(self._arg_descriptors,
-                                           arg_types=GH_VALID_FIELD_NAMES,
-                                           arg_accesses=[AccessType.READ])
+            farg_read = psyGen.args_filter(
+                self._arg_descriptors,
+                arg_types=LFRicArgDescriptor.VALID_FIELD_NAMES,
+                arg_accesses=[AccessType.READ])
             write_accesses = AccessType.all_write_accesses()
-            farg_write = psyGen.args_filter(self._arg_descriptors,
-                                            arg_types=GH_VALID_FIELD_NAMES,
-                                            arg_accesses=write_accesses)
+            farg_write = psyGen.args_filter(
+                self._arg_descriptors,
+                arg_types=LFRicArgDescriptor.VALID_FIELD_NAMES,
+                arg_accesses=write_accesses)
             if len(farg_read) != 1:
                 raise ParseError(
                     "Kernel {0} has a read-only CMA operator. In order "
@@ -1409,7 +871,8 @@ class DynKernMetadata(KernelType):
                 # A valid matrix-matrix kernel must only have CMA operators
                 # and scalars as arguments.
                 scalar_args = psyGen.args_filter(
-                    self._arg_descriptors, arg_types=GH_VALID_SCALAR_NAMES)
+                    self._arg_descriptors,
+                    arg_types=LFRicArgDescriptor.VALID_SCALAR_NAMES)
                 if (len(scalar_args) + len(cwise_ops)) != \
                    len(self._arg_descriptors):
                     raise ParseError(
@@ -3075,8 +2538,8 @@ class DynFields(DynCollection):
         '''
         api_config = Config.get().api_conf("dynamo0.3")
 
-        fld_args = psyGen.args_filter(self._kernel.args,
-                                      arg_types=GH_VALID_FIELD_NAMES)
+        fld_args = psyGen.args_filter(
+            self._kernel.args, arg_types=LFRicArgDescriptor.VALID_FIELD_NAMES)
         for fld in fld_args:
             undf_name = fld.function_space.undf_name
             intent = fld.intent
@@ -3145,7 +2608,7 @@ class LFRicRunTimeChecks(DynCollection):
         existing_checks = []
         for kern_call in self._invoke.schedule.kernels():
             for arg in kern_call.arguments.args:
-                if arg.type not in GH_VALID_FIELD_NAMES:
+                if arg.type not in LFRicArgDescriptor.VALID_FIELD_NAMES:
                     # This check is limited to fields
                     continue
                 fs_name = arg.function_space.orig_name
@@ -3225,7 +2688,8 @@ class LFRicRunTimeChecks(DynCollection):
         modified_fields = []
         for call in self._invoke.schedule.kernels():
             for arg in call.arguments.args:
-                if (arg.text and arg.type in GH_VALID_FIELD_NAMES and
+                if (arg.text and arg.type in
+                        LFRicArgDescriptor.VALID_FIELD_NAMES and
                         arg.access != AccessType.READ and
                         not [entry for entry in modified_fields if
                              entry[0].name == arg.name]):
@@ -3327,7 +2791,7 @@ class DynProxies(DynCollection):
         parent.add(CommentGen(parent, ""))
         for arg in self._invoke.psy_unique_vars:
             # We don't have proxies for scalars
-            if arg.type in GH_VALID_SCALAR_NAMES:
+            if arg.type in LFRicArgDescriptor.VALID_SCALAR_NAMES:
                 continue
             if arg.vector_size > 1:
                 # the range function below returns values from
@@ -3364,7 +2828,7 @@ class DynCellIterators(DynCollection):
             return
         first_var = None
         for var in self._invoke.psy_unique_vars:
-            if var.type not in GH_VALID_SCALAR_NAMES:
+            if var.type not in LFRicArgDescriptor.VALID_SCALAR_NAMES:
                 first_var = var
                 break
         if not first_var:
@@ -3451,15 +2915,16 @@ class DynScalarArgs(DynCollection):
                 self._real_scalars[intent] = []
                 self._int_scalars[intent] = []
             for arg in self._calls[0].arguments.args:
-                if arg.type in GH_VALID_SCALAR_NAMES:
+                if arg.type in LFRicArgDescriptor.VALID_SCALAR_NAMES:
                     if arg.type == "gh_real":
                         self._real_scalars[arg.intent].append(arg.name)
                     elif arg.type == "gh_integer":
                         self._int_scalars[arg.intent].append(arg.name)
                     else:
                         raise InternalError(
-                            "Scalar type '{0}' is in GH_VALID_SCALAR_NAMES but"
-                            " not handled in DynScalarArgs".format(arg.type))
+                            "Scalar type '{0}' is in LFRicArgDescriptor."
+                            "VALID_SCALAR_NAMES but is not handled in "
+                            "DynScalarArgs".format(arg.type))
 
     def _invoke_declarations(self, parent):
         '''
@@ -3794,7 +3259,7 @@ class DynMeshes(object):
         # kernels in this invoke.
         self._first_var = None
         for var in unique_psy_vars:
-            if var.type not in GH_VALID_SCALAR_NAMES:
+            if var.type not in LFRicArgDescriptor.VALID_SCALAR_NAMES:
                 self._first_var = var
                 break
 
@@ -5210,7 +4675,7 @@ class DynInvoke(Invoke):
             # global sum calls
             for loop in self.schedule.loops():
                 for scalar in loop.args_filter(
-                        arg_types=GH_VALID_SCALAR_NAMES,
+                        arg_types=LFRicArgDescriptor.VALID_SCALAR_NAMES,
                         arg_accesses=AccessType.get_valid_reduction_modes(),
                         unique=True):
                     global_sum = DynGlobalSum(scalar, parent=loop.parent)
@@ -5230,11 +4695,11 @@ class DynInvoke(Invoke):
         :raises InternalError: if an invalid access is specified, i.e. \
                 not of type AccessType.
         '''
-        if datatype not in GH_VALID_ARG_TYPE_NAMES:
+        if datatype not in LFRicArgDescriptor.VALID_ARG_TYPE_NAMES:
             raise GenerationError(
                 "unique_proxy_declarations called with an invalid datatype. "
                 "Expected one of '{0}' but found '{1}'".
-                format(str(GH_VALID_ARG_TYPE_NAMES), datatype))
+                format(str(LFRicArgDescriptor.VALID_ARG_TYPE_NAMES), datatype))
         if access and not isinstance(access, AccessType):
             api_config = Config.get().api_conf("dynamo0.3")
             valid_names = api_config.get_valid_accesses_api()
@@ -6611,7 +6076,7 @@ class DynLoop(Loop):
                 self.set_upper_bound("ndofs")
         else:
             if Config.get().distributed_memory:
-                if self._field.type in GH_VALID_OPERATOR_NAMES:
+                if self._field.type in LFRicArgDescriptor.VALID_OPERATOR_NAMES:
                     # We always compute operators redundantly out to the L1
                     # halo
                     self.set_upper_bound("cell_halo", index=1)
@@ -6885,7 +6350,7 @@ class DynLoop(Loop):
                     "currently unsupported for kernels with stencil "
                     "accesses. Found '{0}'.".format(self._upper_bound_name))
             return self._upper_bound_name in ["cell_halo", "ncells"]
-        if arg.type in GH_VALID_SCALAR_NAMES:
+        if arg.type in LFRicArgDescriptor.VALID_SCALAR_NAMES:
             # scalars do not have halos
             return False
         if arg.is_operator:
@@ -7298,9 +6763,9 @@ class DynKern(CodedKern):
                 pre = "iscalar_"
             else:
                 raise GenerationError(
-                    "load_meta expected one of '{0}' but "
-                    "found '{1}'".format(GH_VALID_ARG_TYPE_NAMES,
-                                         descriptor.type))
+                    "load_meta expected one of '{0}' but found '{1}'".
+                    format(LFRicArgDescriptor.VALID_ARG_TYPE_NAMES,
+                           descriptor.type))
             args.append(Arg("variable", pre+str(idx+1)))
 
             if descriptor.stencil:
@@ -7667,9 +7132,9 @@ class DynKern(CodedKern):
         parent_loop = self.parent.parent
 
         # Check whether this kernel reads from an operator
-        op_args = parent_loop.args_filter(arg_types=GH_VALID_OPERATOR_NAMES,
-                                          arg_accesses=[AccessType.READ,
-                                                        AccessType.READWRITE])
+        op_args = parent_loop.args_filter(
+            arg_types=LFRicArgDescriptor.VALID_OPERATOR_NAMES,
+            arg_accesses=[AccessType.READ, AccessType.READWRITE])
         if op_args:
             # It does. We must check that our parent loop does not
             # go beyond the L1 halo.
@@ -7774,14 +7239,14 @@ class ArgOrdering(object):
             # the coarse to the fine mesh.
             self.cell_map()
 
-        # for each argument in the order they are specified in the
+        # For each argument in the order they are specified in the
         # kernel metadata, call particular methods depending on what
         # type of argument we find (field, field vector, operator or
         # scalar). If the argument is a field or field vector and also
         # has a stencil access then also call appropriate stencil
         # methods.
         for arg in self._kern.arguments.args:
-            if arg.type in GH_VALID_FIELD_NAMES:
+            if arg.type in LFRicArgDescriptor.VALID_FIELD_NAMES:
                 if arg.vector_size > 1:
                     self.field_vector(arg)
                 else:
@@ -7801,14 +7266,13 @@ class ArgOrdering(object):
                 self.operator(arg)
             elif arg.type == "gh_columnwise_operator":
                 self.cma_operator(arg)
-            elif arg.type in GH_VALID_SCALAR_NAMES:
+            elif arg.type in LFRicArgDescriptor.VALID_SCALAR_NAMES:
                 self.scalar(arg)
             else:
-                raise GenerationError(
-                    "Unexpected arg type found in dynamo0p3.py:"
-                    "ArgOrdering:generate(). Expected one of '{0}' "
-                    "but found '{1}'".format(GH_VALID_ARG_TYPE_NAMES,
-                                             arg.type))
+                raise InternalError(
+                    "ArgOrdering.generate(): Unexpected argument type found "
+                    "in dynamo0p3.py. Expected one of '{0}' but found '{1}'".
+                    format(LFRicArgDescriptor.VALID_ARG_TYPE_NAMES, arg.type))
         # For each function space (in the order they appear in the
         # metadata arguments)
         for unique_fs in self._kern.arguments.unique_fss:
@@ -8376,11 +7840,12 @@ class KernCallArgList(ArgOrdering):
         farg = self._kern.arguments.get_arg_on_space(fspace)
         # Sanity check - expect the enforce_bc_code kernel to only have
         # a field argument.
-        if farg.type not in GH_VALID_FIELD_NAMES:
+        if farg.type not in LFRicArgDescriptor.VALID_FIELD_NAMES:
             raise GenerationError(
                 "Expected an argument of {0} type from which to look-up "
                 "boundary dofs for kernel {1} but got '{2}'".
-                format(GH_VALID_FIELD_NAMES, self._kern.name, farg.type))
+                format(LFRicArgDescriptor.VALID_FIELD_NAMES,
+                       self._kern.name, farg.type))
         base_name = "boundary_dofs_" + farg.name
         name = self._kern.root.symbol_table.name_from_tag(base_name)
         self._arglist.append(name)
@@ -8721,10 +8186,10 @@ class KernStubArgList(ArgOrdering):
 
         :raises InternalError: if the argument is not a recognised scalar type.
         '''
-        if arg.type not in GH_VALID_SCALAR_NAMES:
+        if arg.type not in LFRicArgDescriptor.VALID_SCALAR_NAMES:
             raise InternalError(
                 "Expected argument type to be one of '{0}' but got '{1}'".
-                format(GH_VALID_SCALAR_NAMES, arg.type))
+                format(LFRicArgDescriptor.VALID_SCALAR_NAMES, arg.type))
         self._arglist.append(arg.name)
 
     def fs_common(self, function_space):
@@ -9316,13 +8781,14 @@ class DynKernelArguments(Arguments):
         of type op_type (either gh_operator [LMA] or gh_columnwise_operator
         [CMA]). If op_type is None then searches for *any* valid operator
         type. '''
-        if op_type and op_type not in GH_VALID_OPERATOR_NAMES:
+        if op_type and op_type not in LFRicArgDescriptor.VALID_OPERATOR_NAMES:
             raise GenerationError(
-                "If supplied, op_type must be a valid operator type (one "
-                "of {0}) but got {1}".format(GH_VALID_OPERATOR_NAMES, op_type))
+                "If supplied, 'op_type' must be a valid operator type (one "
+                "of {0}) but got '{1}'".
+                format(LFRicArgDescriptor.VALID_OPERATOR_NAMES, op_type))
         if not op_type:
             # If no operator type is specified then we match any type
-            op_list = GH_VALID_OPERATOR_NAMES
+            op_list = LFRicArgDescriptor.VALID_OPERATOR_NAMES
         else:
             op_list = [op_type]
         for arg in self._args:
@@ -9358,9 +8824,10 @@ class DynKernelArguments(Arguments):
         # Since we always compute operators out to the L1 halo we first
         # check whether this kernel writes to an operator
         write_accesses = AccessType.all_write_accesses()
-        op_args = psyGen.args_filter(self._args,
-                                     arg_types=GH_VALID_OPERATOR_NAMES,
-                                     arg_accesses=write_accesses)
+        op_args = psyGen.args_filter(
+            self._args,
+            arg_types=LFRicArgDescriptor.VALID_OPERATOR_NAMES,
+            arg_accesses=write_accesses)
         if op_args:
             return op_args[0]
 
@@ -9369,9 +8836,10 @@ class DynKernelArguments(Arguments):
         # we are prolonging (and thus writing to a field on the fine mesh)
         # or restricting.
         if self._parent_call.is_intergrid:
-            fld_args = psyGen.args_filter(self._args,
-                                          arg_types=GH_VALID_FIELD_NAMES,
-                                          arg_meshes=["gh_coarse"])
+            fld_args = psyGen.args_filter(
+                self._args,
+                arg_types=LFRicArgDescriptor.VALID_FIELD_NAMES,
+                arg_meshes=["gh_coarse"])
             return fld_args[0]
 
         # This is not an inter-grid kernel and it does not write to an
@@ -9384,9 +8852,10 @@ class DynKernelArguments(Arguments):
         # a continuous FS is modified then our iteration space must be
         # larger (include L1-halo cells)
         write_accesses = AccessType.all_write_accesses()
-        fld_args = psyGen.args_filter(self._args,
-                                      arg_types=GH_VALID_FIELD_NAMES,
-                                      arg_accesses=write_accesses)
+        fld_args = psyGen.args_filter(
+            self._args,
+            arg_types=LFRicArgDescriptor.VALID_FIELD_NAMES,
+            arg_accesses=write_accesses)
         if fld_args:
             for spaces in [FunctionSpace.CONTINUOUS_FUNCTION_SPACES,
                            FunctionSpace.VALID_ANY_SPACE_NAMES,
@@ -9396,8 +8865,9 @@ class DynKernelArguments(Arguments):
                         return arg
 
         # No modified fields or operators. Check for unmodified fields...
-        fld_args = psyGen.args_filter(self._args,
-                                      arg_types=GH_VALID_FIELD_NAMES)
+        fld_args = psyGen.args_filter(
+            self._args,
+            arg_types=LFRicArgDescriptor.VALID_FIELD_NAMES)
         if fld_args:
             return fld_args[0]
 
@@ -9639,7 +9109,7 @@ class DynKernelArgument(KernelArgument):
                     "associated with this argument (fss='{1}')".format(
                         function_space.orig_name,
                         self.function_space_names))
-        if self._type in GH_VALID_FIELD_NAMES:
+        if self._type in LFRicArgDescriptor.VALID_FIELD_NAMES:
             return "vspace"
         elif self.is_operator:
             if function_space.orig_name == self.descriptor.function_space_from:
@@ -9668,7 +9138,7 @@ class DynKernelArgument(KernelArgument):
     def is_scalar(self):
         ''':return: whether this variable is a scalar variable or not.
         :rtype: bool'''
-        return self.type in GH_VALID_SCALAR_NAMES
+        return self.type in LFRicArgDescriptor.VALID_SCALAR_NAMES
 
     @property
     def mesh(self):
@@ -9835,7 +9305,7 @@ class DynKernelArgument(KernelArgument):
                  False otherwise.
         :rtype: bool
         '''
-        return self._type in GH_VALID_OPERATOR_NAMES
+        return self._type in LFRicArgDescriptor.VALID_OPERATOR_NAMES
 
 
 class DynKernCallFactory(object):
